+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.1.0 - delivery @27/07/2020
+
+- remove fill options
+- update examples and README
+- add tests for debug and Sparkline (calculate and generate)
+
 ## 1.0.0 - delivery @23/07/2020
 
 - first delivery